+++ src/utils/index.ts
@@ -27,7 +27,7 @@ import * as validator from './validator';
  * @param {object} obj The object whose properties to rename.
  * @param {object} keyMap The mapping from old to new property names.
  */
-export function renameProperties(obj: object, keyMap: { [key: string]: string }): void {
+export function renameProperties(obj: {[key: string]: any}, keyMap: { [key: string]: string }): void {
   Object.keys(keyMap).forEach((oldKey) => {
     if (oldKey in obj) {
       const newKey = keyMap[oldKey];
@@ -105,7 +105,9 @@ export function toWebSafeBase64(data: Buffer): string {
 export function formatString(str: string, params?: object): string {
   let formatted = str;
   Object.keys(params || {}).forEach((key) => {
-    formatted = formatted.replace(new RegExp('{' + key + '}', 'g'), params[key]);
+    formatted = formatted.replace(
+        new RegExp('{' + key + '}', 'g'),
+        (params as {[key: string]: string})[key]);
   });
   return formatted;
 }